chore: move types and constants into appropriate files
