+++ src/helpers/constants.ts
@@ -1,0 +1,6 @@
+export const ResponseKeys = ['headers', 'ok', 'status', 'statusText', 'type', 'url', 'redirected'];
+export const AnimalBaseResponseKeys = ['page', 'sort', 'animals'];
+export const BASE_URL = 'https://stapi.co/api';
+export const paths = {
+  animal: '/v1/rest/animal/search',
+};
